composable: refactor validation into composable and utils
