+++ src/components/basic/UInput.vue
@@ -4,7 +4,7 @@
       <span v-if="required"> *</span>
     </label>
     <input :type="type" :value="modelValue"
-           @change="$emit('update:modelValue', $event.target.value); validate($event.target.value);"
+           @change="$emit('update:modelValue', $event.target.value); validate(state, props, $event.target.value);"
            :class="classes"
            :placeholder="placeholder"/>
     <p v-if="!state.valid" v-for="error in state.errors" class="mt-1 text-sm text-red-600" id="error">{{ error }}</p>
@@ -13,6 +13,7 @@
 
 <script setup>
 import {computed, reactive} from 'vue'
+import {validate} from '@/composable/useValidation.js'
 
 const state = reactive({valid: true, errors: []})
 
@@ -49,30 +50,5 @@ const classes = computed(() => {
   }
 })
 
-function validate(value) {
-  state.valid = true
-  state.errors = []
-  if (props.required) {
-    isEmpty(value)
-  }
 
-  if (state.valid && props.type === 'email') {
-    isEMailValid(value)
-  }
-}
-
-function isEmpty(value) {
-  if (value.length === 0) {
-    state.valid = false
-    state.errors.push('This field cannot be empty')
-  }
-}
-
-// TODO externalize in a specific file once more rules are defined
-function isEMailValid(email) {
-  if (!(/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(email))) {
-    state.valid = false
-    state.errors.push('The email is not valid')
-  }
-}
 </script>
